refactor(ingest/kafka-connect): define interface for new connector impl (#12149)
